requirements.txt
Flask
openai
pandas
pyyaml